requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.0
jinja2>=3.1.0
openai-whisper>=20240927
pyannote.audio>=3.1.0
torch>=2.4.0
torchaudio>=2.4.0
python-magic>=0.4.27
setuptools>=75.0.0
